fix: chnage int size from i32 to i64

+++ src/day01/p1.rs
@@ -1,4 +1,4 @@
-pub fn solve(inp: Vec<&str>) -> Result<i32, Box<dyn std::error::Error>> {
+pub fn solve(inp: Vec<&str>) -> Result<i64, Box<dyn std::error::Error>> {
     let mut prev = -1;
     let mut counter = 0;
     for raw_num in inp {
